ve.ce.Surface: Remove font-size: 1em;
In the MediaWiki integration we want the surface to look the same
as in read mode. For that, it we would apply the same class (e.g.
mw-body-content). However, that class assumes its font-size in em
size to be relative to #content. It doesn't have to be a direct
child of #content (the extra nestings Surface adds are fine), but
our contains between #content and surface must not change the
font-size as that will affect the font-size of the content.

Change-Id: Ibcee8e6ee8356fa6c2575dd6203c5d1e3ea78e63